fix: resolve compiler warnings in clean build - part 2
- socket.c: Fixed socket macros to use intptr_t cast for proper int-to-void* conversion
- mixer.c: Simplified source validation check removing pointer-to-int comparison
- grep.c: Removed unused skip_ansi_codes function
- splash.c: Fixed sign comparison in elapsed_ms calculation
- colorscheme.c: Fixed struct initialization with proper braces
- handlers.c: Kept unused parameter desc as required by function pointer type
- keyboard_handler.c: Cast key to int in switch statement
- packet.c: Changed loop counter from int to size_t for sign compatibility
- websocket/transport.c: Removed unused CHUNK_SIZE constant
- websocket/server.c: Removed unused websocket_extensions array

Reduced warnings from 29 to 14 remaining (mostly struct field initializers).

+++ lib/network/packet.c
@@ -601,7 +601,7 @@ packet_recv_result_t receive_packet_secure(socket_t sockfd, void *crypto_ctx, bo
 
   // DEBUG: Log raw header bytes for debugging packet misalignment
   char header_hex[95];
-  for (int i = 0; i < sizeof(header); i++) {
+  for (size_t i = 0; i < sizeof(header); i++) {
     sprintf(&header_hex[i * 2], "%02x", ((uint8_t *)&header)[i]);
   }
   header_hex[94] = '\0';
